Implement ADS client (#12)
This change includes the new `diderot.ADSClient` type, which is a
generic ADS client that can be used to talk to any ADS server (including
notably the `diderot.ADSServer` :)). It also includes a number of
improvements to the way glob collection URLs are handled, and makes it
easier to manage them:
1. The parsing and creation methods for glob collection URLs now accept
type parameters instead of the type string. This makes it less error
prone and more type-safe.
2. There is now a function to create an `ads.GlobCollectionURL`:
`ads.NewGlobCollectionURL()`. Along with the new method `MemberURN`, it
makes it easier to create glob collection URNs.
3. `ExtractGlobCollectionURLFromResourceURN` was renamed to
`ParseGlobCollectionURN` and now returns the parsed resource name, i.e.
the last element of the glob collection. This makes it easier to find
out whether a given URI is a URN (i.e. a specific resource within the
collection) or the URL for that collection (i.e. ends with `/*`)