Fix race condition in glob collections (#10)
When subscribing to a glob collection that is being updated, the
`nonNilValueNames` set was being accessed in an unprotected way. This
was generally safe as it was only used by new subscribers when they
subscribed to empty collections. However, it was still potentially
problematic. This fix not only addresses this race condition, but by
refactoring `NotifyHandlerAfterSubscription`, it also allows multiple
subscribers to subscribe to the glob collection concurrently. This will
hopefully speed up the processing of new clients subscribing to many
large glob collections all at once (which can happen during startup).