Correctly handle deletion of cache entry (#17)
See documentation in code for more details, but the high-level problem
is that an entry was deleted too early from the cache, orphaning the
notification loop. If the same resource gets re-created, it would create
a new entry and execute a new notification loop. These two loops would
then deliver competing notifiations to subscribers.